requirements.txt
# JohnnyBets API Dependencies (Production)

# LangGraph/LangChain
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain>=0.3.0

# Web Framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
sse-starlette>=2.0.0

# Data Processing
pandas>=2.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.27.0

# Environment
python-dotenv>=1.0.0

# Validation
pydantic>=2.0.0

# Kalshi API
kalshi-python>=1.0.0
cryptography>=41.0.0
